CSS & mobile fixes.
* Added a link to each comment.  Closes #6 .
* Main thread scrollbar no longer overlaps top menu.  Closes #5.
* Added basic support for mobile devices.
* Fixed an issue with server-side date string generation.  Previously it
would show time instead of date if the time was within the last 24
hours, which could be confusing.  Changed ot only show time if the
datetime value was for the current day.
* Added an option to enable/disable notification under user settings.
Actual functionality is not implented yet.
* Reorganized some of the Thread.js code, moving more generic stuff int
util.js.
* Removed Sugar.js library as it was only used for some very minor stuff
which I re-implemented in Util.js instead.
* Removed all references to two factor and phone authentication, we
don't need that.