Ton of fixes and features.
The threads list now uses an infinite scroller, loading on demand.  Only
loads 10 threads initially.  Closes #11
Links to tweets are now converted to embedded tweets automatically.
Lazy loading of images as they scroll into view via Semantic UI except
animated gifs because
Animated gifs no longer play by default, have to click the play button
to run them.  Clicking again pauses them.  Uses Gifffer
Threw in timeago to make datetime display work better
The unread items indicator is now clickable on mobile and browse to the
next unread comment.
Mobile only closes the thread list on start if you have a thread
selected now.
Added the Authorize attribute to the SignalrHub which should fix the
shitty problem with signalr trying to reconnect over and over but
throwing an exception because I was stupid. Fixes #10
Trailing spaces at the end of a line after an unlinked URL no-longer
breaks the autolinked URL. Fixes #17
Threads loaded by clicking on the thread list are now loaded via SignalR
instead of the REST API.  That seems to be much faster.
Fixed code formatter so it actually displays properly after you save the
comment or reload the page.
Got rid of the stupid CustomHtmlHelperService class.  The linkify code
was moved into the Comment class and runs whenever Comment.Text is
modified.
Added a Comment.DisplayText property that applies some
formatting/filtering to comments for output to the page without
modifying the data that's saved.  Decided to go this route for shit like
adding the necessary classes for lazy loading images and the gif player
so next time shit has to be changed around we don't have to worry about
updating the content in the DB.
Moved the SiteRoot config value to a static readonly property on the
Startup class because why not?  Making that available via a service and
using DI was a PITA for certain things, mainly the automapper, so I put
it there since it's not going to change.
The old CustomHtmlHelperService class is now the stubble/mustache
renderer service and is a singelton service, but not a singleton itself
since it only ever needs to be created once in startup when it's added
as a singleton service instance.  It precompiles and caches the mustache
templates and is used by the controllers and signlar to render content.
Fixed an issue where new user creation wasn't subscribing the new user
to all threads and currently everyone needs to be or it breaks shit when
people add new comments to threads you're not subscribed to.  Eventually
I'll finish the subscription functionality and fix all that.

Probably broke a whole bunch of crap.  Yay.