Thread loading optimization.
Disabled the events for the thred list infinite scroller while loading
threads and applied lazy loading, gifffer and twitter embedding to a
document fragment of a new thread before replaceing the current thread
with the new one, so the thread only renders once.  Sped up load times
quite a bit.
Added a loader that displays over the thread div when loading a new
thread.